\name{get_miRNA_family_info}
\alias{get_miRNA_family_info}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
MicroRNA family information
}
\arguments{
  \item{datapath}{data path
}
}
\description{
Get human miRNA family information obtained frin TargetScanHuman 6 website
}
\usage{
get_miRNA_family_info(datapath = system.file("extdata/miR_Family_Info.txt", package = "TargetScoreData"))
}
