\name{get_TargetScanHuman_contextScore}
\alias{get_TargetScanHuman_contextScore}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
get_TargetScanHuman_contextScore
}
\arguments{
  \item{datapath}{data path
}
}
\description{
TargetScan context score for all of the predicted sites (including conserved and nonconserved sites) downloaded from TargetScan 6 website (http://www.targetscan.org/cgi-bin/targetscan/data_download.cgi?db=vert_61)
}
\usage{
get_TargetScanHuman_contextScore(datapath = system.file("extdata/TargetScanHuman_contextScore.RData", package = "TargetScoreData"))
}
