\name{get_TargetScanHuman_PCT}
\alias{get_TargetScanHuman_PCT}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
get_TargetScanHuman_PCT
}
\arguments{
  \item{datapath}{data path
}
}
\description{
TargetScan PCT for all of the predicted sites (including conserved and nonconserved sites) downloaded from TargetScan 6 website (http://www.targetscan.org/cgi-bin/targetscan/data_download.cgi?db=vert_61)
}
\usage{
get_TargetScanHuman_PCT(datapath = system.file("extdata/TargetScanHuman_PCT.RData", package = "TargetScoreData"))
}
