\name{get_validated_targets}
\alias{get_validated_targets}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
get_validated_targets
}
\arguments{
  \item{datapath}{data path
}
}
\description{
Get validated targets of human miRNA obtained from mirTarBase v3.5
}
\usage{
get_validated_targets(datapath = system.file("extdata/hsa_MTI.xls", package = "TargetScoreData"))
}
